validation for description
